requirements.txt
aiohttp
asyncio
SQLAlchemy
python-dotenv